+++ DQB2_WorldEdit.csproj
@@ -2,5 +2,6 @@
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
     <EnableDynamicLoading>true</EnableDynamicLoading>
+    <RootNamespace>EyeOfRubiss</RootNamespace>
   </PropertyGroup>
 </Project>
