+++ Integration/IntegrationMain.cs
@@ -9,7 +9,7 @@ public partial class IntegrationMain : Control
 {
 	private IntegrationWorldEditor WorldEditor;
 	private IWorld world;
-	private IDriver driver;
+	private IDriver driver; // needed to make sure the driver doesn't get GC'd which would cause the FileSystemWatcher to stop firing
 
 	private static string DriverFile { get; set; }
 
@@ -17,7 +17,7 @@ public partial class IntegrationMain : Control
 	/// Returns true if the command line args contain something like `--driverFile C:\Temp\Blah\driver.json`
 	/// telling us to run in integration mode.
 	/// </summary>
-	public static bool IsRunningIntegrationMode(IReadOnlyList<string> cmdLineArgs)
+	public static bool ShouldSwitchToIntegrationMode(IReadOnlyList<string> cmdLineArgs)
 	{
 		var args = cmdLineArgs.Select(x => x?.ToLowerInvariant()).ToList();
 		var idx = args.IndexOf("--driverfile");
@@ -40,15 +40,27 @@ public partial class IntegrationMain : Control
 
 		if (string.IsNullOrEmpty(DriverFile))
 		{
-			throw new Exception($"Assert fail: DriverFile should have been set by {nameof(IsRunningIntegrationMode)}");
+			throw new Exception($"Assert fail: DriverFile should have been set by {nameof(ShouldSwitchToIntegrationMode)}");
 		}
 
 		GD.Print($"{nameof(IntegrationMain)} has loaded... Now looking at {DriverFile}");
-		driver = EyeOfRubiss.Integration.FSWatcher.FileSystemDriver.Create(DriverFile);
+
+		IDriver driver;
+		string json = System.IO.File.ReadAllText(DriverFile);
+		var peek = System.Text.Json.JsonSerializer.Deserialize<DriverFileContentBase>(json);
+		if (peek.IntegrationType == IntegrationTypeConstants.FSWatcher)
+		{
+			driver = EyeOfRubiss.Integration.FSWatcher.FileSystemDriver.Create(DriverFile);
+		}
+		else
+		{
+			throw new BadIntegrationException($"unrecognized IntegrationType given: {peek.IntegrationType}");
+		}
+
+		this.driver = driver;
 		driver.WorldUpdated += Driver_WorldUpdated;
 		world = driver.World;
 		GD.Print("Driver created!");
-
 		WorldEditor.LoadWorld(world);
 	}
